remove old code
- files for old test
- organized constants file
- added an __init__.py for apis
