+++ src/constants.py
@@ -1,18 +1,24 @@
 """
 All Constants that are used throughout the project
 """
-ADDRESS = "0x9008D19f58AAbD9eD0D60971565AA8510560ab41"
-# ADDRESS is deprecated, use SETTLEMENT_CONTRACT_ADDRESS instead
-SETTLEMENT_CONTRACT_ADDRESS = "0x9008D19f58AAbD9eD0D60971565AA8510560ab41"
+
+# main loop
+SLEEP_TIME_IN_SEC = 10
+
+# surplus tests
 ABSOLUTE_ETH_FLAG_AMOUNT = 0.002
 REL_DEVIATION_FLAG_PERCENT = 0.1
+
 # cost coverage test
 COST_COVERAGE_ABSOLUTE_DEVIATION_ETH = 0.005
 COST_COVERAGE_RELATIVE_DEVIATION = 0.20
+
 # fee quote test
 FEE_ABSOLUTE_DEVIATION_ETH_FLAG = 0.002
 FEE_RELATIVE_DEVIATION_FLAG = 0.20
-SLEEP_TIME_IN_SEC = 10
+
+# requests
+SETTLEMENT_CONTRACT_ADDRESS = "0x9008D19f58AAbD9eD0D60971565AA8510560ab41"
 REQUEST_TIMEOUT = 5
 SUCCESS_CODE = 200
 FAIL_CODE = 404
